+++ crates/primitives/src/lib.rs
@@ -7,7 +7,6 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(feature = "nightly", feature(hasher_prefixfree_extras))]
 #![cfg_attr(feature = "nightly", feature(core_intrinsics))]
-#![cfg_attr(feature = "nightly", feature(likely_unlikely))]
 #![cfg_attr(feature = "nightly", allow(internal_features))]
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
@@ -16,8 +15,8 @@ extern crate alloc;
 
 use paste as _;
 use rustc_hash as _;
-#[cfg(feature = "sha3-keccak")]
 use sha3 as _;
+#[cfg(feature = "tiny-keccak")]
 use tiny_keccak as _;
 
 #[cfg(feature = "postgres")]
@@ -41,7 +40,8 @@ pub use aliases::{
 mod bits;
 pub use bits::{
     Address, AddressChecksumBuffer, AddressError, BLOOM_BITS_PER_ITEM, BLOOM_SIZE_BITS,
-    BLOOM_SIZE_BYTES, Bloom, BloomInput, FixedBytes, Function,
+    BLOOM_SIZE_BYTES, Bloom, BloomInput, FixedBytes, FixedBytesSliceExt, FixedBytesVecExt,
+    Function,
 };
 #[cfg(feature = "rkyv")]
 pub use bits::{
@@ -93,7 +93,7 @@ pub use ::hex::serde as serde_hex;
 // Not public API.
 #[doc(hidden)]
 pub mod private {
-    pub use alloc::vec::Vec;
+    pub use alloc::{borrow::Cow, vec::Vec};
     pub use core::{
         self,
         borrow::{Borrow, BorrowMut},
@@ -128,4 +128,7 @@ pub mod private {
 
     #[cfg(feature = "sqlx")]
     pub use sqlx_core;
+
+    #[cfg(feature = "schemars")]
+    pub use schemars;
 }